Parser overhaul
This is the first commit in the parser rewrite.

The main change is that the old comment representation, which was
excessively permissive, has been replaced by a more restrictive,
precise, and therefore informative representation. This makes the HTML
generator, and several other bits of code, much simpler, and will also
make it easy to write additional generators in the future.

Correspondingly, the former Octavius has been deleted, and replaced with
a new parser that produces this representation.

The new parser comes with (currently) 459 test cases for various syntax
corner cases.

Several features have been temporarily disabled in the surface syntax,
though the internal code for handling them is still present in odoc.