Restore @canonical tag
This is an extension to the ocamldoc syntax, apparently used by
Jbuilder. Jbuilder adds attributes to files it has processed containing
the @canonical tag. This apparently has to do with wrapped modules in
libraries, and hiding their internal names.

I would prefer to add a new kind of attribute instead of a new concrete
syntax, unless there is a clear human use case for this.