Restore module list support
The support is partial for now. It appears module name resolution was
partly bugged before the parser overhaul anyway. As a result, summary
text was not being displayed. Until that resolution bug is fixed, the
code has no provision for displaying summary text, as there is no way to
test that the text is correct anyway.