Check for bad heading levels in only one place
Before this commit, there was a lexer error rule for multi-digit heading
levels, which flagged section levels that are too large. Separately, a
check in semantics.ml checked for single-digit heading levels that are
too large. The lexer rule that accepted heading levels, for further
checking by semantics.ml, accepted only single digits.

This commit relaxes the lexer rule to accept multi-digit heading levels,
eliminating the error rule for multi-digit heading levels in the lexer.
Instead, all too-large heading levels are caught in semantics.ml. There
is now only one place where the corresponding error is raised. Also, the
check in semantics.ml produces a recoverable warning, while the rule in
the lexer was a fatal error, and would need extra work to be made
recoverable.

Reported locations have changed for some tests, because the location of
the entire heading is reported, rather than just the heading start
token.