Add a child reference type
This will be used in mld files in the following way:

    {!child:childref}

where 'childref' is, for example, 'foo', if you intend to compile an mld file
'foo.mld' as a child of this document, or 'Sexplib0' if you wish the module
'Sexplib0' to be a child document.

Resolution of this reference type is not implemented in this commit.

Signed-off-by: Jon Ludlam <jon@recoil.org>