Add Odoc_parser.position_of_point
This adds a bit of symmetry to the interface. When parsing a comment,
we initially pass in a `Lexing.position` to indicate the start of the
comment. However, the locations within the AST are all represented
as a locally defined type, Loc.point. With the new function we can
turn these values back into Lexing.position values.

This requires storing the initial position value in [t], which is an
internal detail it's not useful to expose. Consequently the type
`Odoc_parser.t` has been made opaque and accessor functions made
available for the warnings and the ast.