Fix default alignment
This commit fixes several alignment issues.

1. No alignment in a cell was encoded as an alignment to `Center.

"Centering" was certainly a typo, but there is also the issue that "no specified
alignment" should really be different than "specified to ..."

For instance (assuming `{R ...}` exists):
```
{R
  {t
    |aaa|bbb|
    |---|---|
    |xxx|yyy|}}
```
would not be right-aligned inside the table, with the previous behaviour.

This commit allows for "unspecified alignment" using an option.

2. No alignment in a table was encoded as an empty list of alignment.

This was wrong, since at some point we might want to raise warnings in case the
number of alignment and the number of columns differ.

Signed-off-by: Paul-Elliot <peada@free.fr>